WIP: tool call divs
